fix: organize files
